refactored some names, to make more clear
