+++ tests/pic_importer_test_class.py
@@ -1,6 +1,6 @@
 """test case of the PicturaeImporter class which runs a reduced init method to use in unittests"""
 import logging
-from tests.casbotany_sql_lite import SqlLiteTools
+from tests.sqlite_csv_utils import SqlLiteTools
 from image_client.importer import Importer
 from image_client.picturae_importer import PicturaeImporter
 from image_client import picturae_config
@@ -10,5 +10,5 @@ class TestPicturaeImporter(PicturaeImporter):
     def __init__(self, date_string, paths):
         Importer.__init__(self, db_config_class=picturae_config, collection_name="Botany")
         self.init_all_vars(date_string=date_string, paths=paths)
-        self.sqllite_csv_tools = SqlLiteTools()
+        self.sqlite_csv_tools = SqlLiteTools()
         self.logger = logging.getLogger("TestPicturaeImporter")
